Set server_addr in audit events from connection info
This sets a useful server IP, when no advertise_ip is set. Previously,
the address was taken from the listener, and is usually "0.0.0.0:3022"
or "[::]:3022".

Also, make the new etcd test skip when etcd isn't running locally (like
with other tests).
Also, add some test cases in utils for IPv6 handling.